added prop pixelRatio; added standalone functions; did some code refactor
